move enum.hpp
