+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/output.hpp
@@ -11,7 +11,7 @@
 
 #include "meta_data.hpp"
 
-#include "../enum.hpp"
+#include "../common/enum.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
