+++ power_grid_model_c/power_grid_model/include/power_grid_model/exception.hpp
@@ -6,8 +6,8 @@
 #ifndef POWER_GRID_MODEL_EXCEPTION_HPP
 #define POWER_GRID_MODEL_EXCEPTION_HPP
 
+#include "common/enum.hpp"
 #include "common/power_grid_model.hpp"
-#include "enum.hpp"
 
 #include <exception>
 #include <string>
