+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/short_circuit_solver.hpp
@@ -11,7 +11,7 @@
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
-#include "../enum.hpp"
+#include "../common/enum.hpp"
 #include "../exception.hpp"
 
 namespace power_grid_model::math_solver {
